@mattbostock 2018-03-21
On closer inspection it looks like we used to use a slice and changed to maps in dab8af3af48f705feb0b7b683b385ffc11ccb3ca.

@beorn7: What do you think about reverting dab8af3af?