@beorn7 2018-03-21
Great stuff.
I guess I only changed to the map because I liked the interface more, and missed that that anonymous function is actually in the hot path.

I prefer your PR over reverting dab8af3af48f705feb0b7b683b385ffc11ccb3ca as it keeps the (presumably) nicer interface but still gets all the performance improvement.